+++ internal/core/domain/transaction_repository.go
@@ -39,7 +39,7 @@ type TransactionRepository interface {
 	// Transaction identified by the given txid.
 	// Generates a TransactionConfirmed event if successfull.
 	ConfirmTransaction(
-		ctx context.Context, txid, blockHash string, blockheight uint32,
+		ctx context.Context, txid, blockHash string, blockheight uint64,
 	) (bool, error)
 	// GetTransaction returns the Transaction identified by the given txid.
 	GetTransaction(ctx context.Context, txid string) (*Transaction, error)
